+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MINOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Predefined "Prometheus" datasource added;  
+* Custom datasource factory methods implemented: `graphiteDatasource(name: String)`,
+`prometheusDatasource(name: String)`,`zabbixDatasource(name: String)`;
+* 
 ### NEXT_VERSION_DESCRIPTION_END
 ## [3.8.1](https://github.com/yoomoney/grafana-dashboard-dsl/pull/20) (07-09-2021)